# backend/src/config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "PwndocNG API Documentation",
    "description": "",
    "version": "1.0.0"
  },
  "host": "/",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/users/refreshtoken": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/users/token": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "totpToken": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "firstname": {
                  "example": "any"
                },
                "lastname": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/api/users/me": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "put": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "currentPassword": {
                  "example": "any"
                },
                "newPassword": {
                  "example": "any"
                },
                "confirmPassword": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "firstname": {
                  "example": "any"
                },
                "lastname": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/users/totp": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "totpToken": {
                  "example": "any"
                },
                "totpSecret": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "totpToken": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/users/{username}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/users/{id}": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "firstname": {
                  "example": "any"
                },
                "lastname": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                },
                "totpEnabled": {
                  "example": "any"
                },
                "enabled": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
  }
}